Time: 14 ms (67.71%), Space: 44.7 MB (54.79%) - LeetHub

+++ 1567-maximum-number-of-vowels-in-a-substring-of-given-length/1567-maximum-number-of-vowels-in-a-substring-of-given-length.java
@@ -10,7 +10,7 @@ class Solution {
             if (i >= k && isVowel(s.charAt(i - k))) {
                 current--;
             }
-            answer = (int) Math.max(current, answer);
+            answer = Math.max(current, answer);
         }
 
         return answer;
